Move plurals helper in their own helper module
