+++ app/helpers/waste_exemptions_engine/application_helper.rb
@@ -13,10 +13,6 @@ module WasteExemptionsEngine
       "#{first_name} #{last_name}"
     end
 
-    def emails_plural(form)
-      form.applicant_email == form.contact_email ? "one" : "many"
-    end
-
     def current_git_commit
       @current_git_commit ||= begin
         sha =
