+++ app/helpers/waste_exemptions_engine/plurals_helper.rb
@@ -1,7 +1,11 @@
 # frozen_string_literal: true
 
 module WasteExemptionsEngine
-  module RegistrationCompleteFormsHelper
+  module PluralsHelper
+    def emails_plural(form)
+      form.applicant_email == form.contact_email ? "one" : "many"
+    end
+
     def exemptions_plural(form)
       form.exemptions.length > 1 ? "many" : "one"
     end
